Enabled the 'Devices' of LEA's peer configurator.
LEA's peer configurator is now able to display all devices of a peer. The devices can be edited (change name, interface etc.) but it is currently not possible to add/remove devices to/from a peer.

Additionally enhanced the `Initialized` component by navigating the user to the error page in case of failures.