Polished some smaller UI parts of LEA.
Added an undeploy action to the `ClusterOverview`.

The clusters card shown by the dashboard does now display the number of deployed and undeployed clusters.

The peers card offers now a button to quickly create a new peer as the clusters card provides this for clusters.

The peer overview's table has now a name column instead of an id column. The same change has been made to the cluster overview to display a clusters name instead of its id.