CARL Persistence -> Introduce proper transactions.
Transactions now span the ResourcesManager-APIs used for modifying persistence:
- `.insert()`
- `.remove()`
- `.resources_mut()`

In particular `.resources_mut()` allows developers to group multiple
requests into a transaction. To achieve this, the API was changed to
require returning a `Result<T, E>` out of the closure. If an `Err` is
returned this way, then everything is rolled back.

Transactions are not supported for in-memory storage, as this would
require implementing a custom transaction mechanism.
This can be problematic when the database state is rolled back, but the
in-memory state is not, since some Resources are always kept
in-memory.
This change should still be an improvement, since when the database
state is not impacted, CARL can be restarted to clear out potentially
inconsistent in-memory state.