CARL Resources -> Initial changes for replacing Postgres with a key-value store.
We found that a key-value store matches our architecture (in particular the ResourceManager) much better.

Within our application, data is transmitted in full, for example when a
device is removed from a PeerDescriptor, then the whole PeerDescriptor
is submitted by the UI. Relational databases, on the other hand, work
best with relative changes and by splitting data up into small tables.

This resulted in the PeerDescriptor being split across 7 database tables
with fairly complex logic to update these correctly when a whole
PeerDescriptor is submitted. In particular, when lists in the
PeerDescriptor had elements removed, we had to perform diffing logic
to figure out what DELETE command to issue to the database.
We had similar problems with the ClusterConfiguration.

Because we retrofitted the relational database, we were not making
use of its querying features anyways. Everything had to be handled
on top of the ResourceManager, which narrowed down the API to be
effectively that of a key-value store.