LEA -> Fix sorting of device selector in cluster overview + code cleanup.
Part of the code cleanup eliminated a crash we observed,
where a device had an interface which was not associated with any peer.
This should never happen, so eliminating the crash might not necessarily
be the best approach, but it will become visible in the UI at least and
prevents the UI from becoming unusable.