Remove explicit configuration of Minimum Supported Rust Version.
We had introduced this to try to keep the dev environments in sync.
Now, we are using `rust-toolchain.toml` for that.

Setting a MSRV without ensuring that it is correct like we did
is not helpful and can be problematic, since the compiler will not report
warnings that would still be valid code in these older Rust versions.