Revert "Doc: Only publish when `main`-branch is updated."
This reverts commit 041d398b962947ec7d32cad5bed4609863e21623.

Now that v0.8 is released, we don't expect major changes to the
documentation in the near future anymore, so there is less of a chance
of it not being applicable to v0.8 after making development changes.

Allowing early updates to the documentation lets us introduce fixes
and smaller changes without having to wait until a release happens.