cargo ci: Introduce `cargo carl-viper` and `cargo lea-viper`.
These commands run CARL and LEA with `--features=viper`.

This needed to be separate from `cargo carl` and `cargo lea`, because these commands directly passthrough everything to the started applications (which we do not want to change, particularly because `cargo cleo` becomes cumbersome to use, for which we initially introduced these shorthands).