EDGAR+CLEO: Introduce mTLS client authentication for connection to CARL.
A larger cleanup of how we load certificates was necessary.
Now the certificates are consistently loaded in the same way, allowing
the default config key to be extended by `.content` to pass certificate
text directly. This is useful in environment variables in particular.

This does not yet cover the OpenTelemetry connection and the NetBird
client is not yet configured correctly either.